requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1
qrcode==7.4.2
weasyprint==62.3
